fix: restore the AI-readiness gate to zero errors
Two unrelated violations had landed together and were failing every PR in
the repository regardless of its contents:

- `abicheck/mcp_server.py` had grown to 2017 lines, over the gate's *hard*
  2000-line cap, which has no allowlist by design.
- `docs/use/output-formats.md`'s example still showed
  `"report_schema_version": "2.26"` after the source of truth moved to 2.27.

The `ai-readiness` CI job stops before its own `pip install -e ".[dev]"`
step when the gate fails, and the per-tier accuracy step that follows
carries `if: always()` -- so it then died on `ModuleNotFoundError: No module
named 'elftools'`. That was a consequence of the first failure, not a third
problem, and clears with it.

The verdict/exit-code/change-entry helpers move verbatim to a new leaf
module, `abicheck/mcp_server_verdicts.py`, re-exported from `mcp_server` so
any existing import path keeps working. Behaviour is unchanged. The module
imports only `checker_policy` and `mcp_shared`'s logger at module scope --
the heavier `severity`/`reporter`/`appcompat` dependencies stay
function-local exactly as they already were -- so the edge from `mcp_server`
is one-directional and no new import cycle forms (`import-cycle-growth`
verified clean).

`scripts/check_ai_readiness.py` now reports 0 errors, down from 2.
