+++ docs/use/output-formats.md
@@ -578,7 +578,7 @@ Every JSON report carries a top-level `report_schema_version` field
 
 ```json
 {
-  "report_schema_version": "2.26",
+  "report_schema_version": "2.27",
   "library": "libfoo.so.1",
   "verdict": "BREAKING"
 }
